moved resource keys to new resource package
